requirements.txt
# Core dependencies
aiohttp==3.9.1
asyncio==3.4.3
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2

# AI and NLP
openai==1.6.1
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
tenacity==8.2.3
unidecode==1.3.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.8.0